Remove NMS dependency, use jitpack for Vault dependency
These changes allow BetterEconomy to be built from source without
requiring the VaultAPI or spigot artifacts to be present locally. This
would help anyone contributing to BetterEconomy, as well as CI.

NMS is not necessary for an economy plugin. If it seriously needs to be
re-added later, it can be. The VaultAPI uses jitpack, and there are
several issues on the Vault repo explaining this.